test(vector_score_source): isolate VecSim timeout tests in their own binary
Address review feedback (MOD-16948): the previous Mutex only serialized tests
within source_pytest_parity, but tests/integration/main.rs compiles policy,
source, and source_pytest_parity into ONE `integration` binary that shares the
process-global VecSim timeout callback. The `source` module's query tests never
took the lock, so under `cargo llvm-cov test` a `*_propagates_timeout` test
could still install `always_timed_out` while a sibling `source` query ran — the
coverage-lane race remained.

Move the two `*_propagates_timeout` tests (and the MockTimeout / callback FFI
helpers) into a dedicated `tests/vecsim_timeout.rs` test binary. Cargo compiles
it as a separate process, so the global-callback mutation can never reach any
query test, on any harness — this makes explicit and harness-independent the
process isolation the design previously leaned on nextest for. The query tests
in source_pytest_parity therefore no longer need any lock; that file's tests are
now byte-identical to before the fix (minus the moved section). The two timeout
tests are serialized against each other within their binary so one's teardown
cannot clear the callback the other is asserting on.
